move more constructor tests, nlargestsmallest, squeeze
